requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b
cachetools==4.1.1
certifi==2020.6.20
celery==4.4.7
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
django-cors-headers==3.5.0
django==3.1.14
django-ckeditor==6.2.0
django-leaflet==0.28.2
django-phonenumber-field[phonenumbers]==6.0.0
djangorestframework==3.13.1
djangorestframework-gis==0.18
djangorestframework-stubs==0.4.2
djangorestframework-jwt==1.11.0
django-celery-beat==2.0.0
django-constance==2.8.0
django-filter==2.4.0
django-picklefield==3.0.1
django-admin-interface==0.18.6
django-postgres-metrics==0.11.0
django-adminactions==1.15
drf-yasg==1.20.0
drf-extra-fields==3.2.1
filelock==3.0.12
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
googlemaps==4.5.3
gunicorn==20.0.4
identify==1.4.30
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
jinja2==2.11.3
markupsafe==1.1.1
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.4
pathspec==0.8.0
pip-tools==5.3.1
Pillow==9.0.0
pre-commit==2.16.0
protobuf==3.15.0
psycopg2-binary==2.8.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2021.3
pyyaml==5.4
redis==3.5.3
regex==2020.7.14
requests==2.27.1
rsa==4.7
ruamel.yaml.clib==0.2.2
ruamel.yaml==0.16.12
sendgrid==6.9.3
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
typed-ast==1.4.1
typing-extensions>=3.10
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.3.0
virtualenv==20.0.31
zipp==3.1.0
